Changed references to reflect new file names
Since I changed the names of some of my modules, I needed to update the
import calls in the wrapper scripts. I also changed the DRUGSdb variable
from an environment variable containing a path name to a hard coded path
name to make it easier for GitHub users to try the scripts.